remove inline rollup hooks on the api (#1027)
Partially reverts #525. This removes the plumbing for the ids
and rollups via the api. The usage for the internal DB have
been retained.

The reason for this is it adds additional complexity and risk
to the backends and for most use-cases we are moving to the
simple aggregator cluster instead (see atlas-aggregator in
iep-apps). As we are not planning to exercise this as scale
anytime soon, it is being removed from the APIs before 1.6
release.