expire interned queries after 12h (#1030)
Switches from a ConcurrentHashMap to a Cache for the interned
queries. This avoids a slow leak over time with churn in the
set of queries. In practice, we have not seen enough churn for
it to matter, but now queries that are no longer in use will
expire and go away.

Fixes #729.