detect empty state and stop tracking it
Updates the online algorithms used with stateful operators
to detect if the state is effectively empty. In that case
the state will be dropped from the state map maintained
during evaluation. This is useful to avoid leaking memory
for tracking the state during long running streaming
evaluations.

There are currently three exceptions that have an unbounded
window and never become empty: ignore-N, sdes, and integral.
Integral is generally not very useful in a streaming context
so we may just prohibit that operation. For sdes (and ignore-N
which is typically only used to align sdes), the current uses
do not involve group by operations, so it shouldn't be an
issue for now. We'll revisit in the future when there are
concrete use-cases that are problematic for that operation.