fix eval state for sparse lines with group by (#1035)
Before, if a line was sparse, for example an error counter
that has a lot of gaps, the state would be handled incorrectly.
There would be one state item per matching grouping when data
was present and they would only get moved forward if there was
data for that key in a given interval. For intervals with no
data, a no data line would get pushed through and have its own
independent state.

With this change, an empty group by result will push an empty
data set through eval and it will now move the state buffers
to allow for expiration and correct handling of subsequent
intervals with data. If there is no resulting data lines from
the overall expression, then a no data line for the overall
expression will be emitted to the user.