add NameValueLengthRule
Allows different limits to be used for the value of the
name key compared to other keys. This is mostly to remove
a difference between the internal Atlas setup at Netflix
and OSS. Internally for historical reasons the value for
the name key is allowed to be longer than other values.
With this change there is no longer a need to have custom
rules.