switch to max normalization for gauges
For use-cases where we perform inline aggregation this
gives the expected value and avoids the need for an
additional disambiguation tag for the aggregator instance.
This also means that if no custom rollups are done the sum
of the gauge values would match expectations rather than
being a multiple of the number of aggregator instances.