add options to tune request handler (#1100)
There are options now to disable automatic handling
of compression and whether or not to setup access
logging. For high volume services the automatic
compression results in a lot of allocations for
intermediate ByteStrings. The json directives will
not automatically wrap with a GZIPInputStream if
the data is GZIP compressed. For some services this
drastically reduces the allocation rate.