optimize ByteStringInputStream for large payloads
Uses the `asByteBuffer` conversion to allow for more
efficient reads. The `copyToArray` was really slow and
was worse than just always using `apply` and reading
a single byte. The byte buffer approach avoids the
intermediate array allocation and supports a fast copy
to another array. For the sample benchmark it is about
10x faster.