avoid delay on finish for local eval processor (#1106)
When using the eval processor for local data, it should
stop quickly after the upstream is complete. Before the
throttling on the Eureka lookup would cause about a 30s
delay.

This changes the Eureka lookup to use an empty source
if there are no remote data sources. Also fixes a bug
with TimeGrouped where it could potentially push without
a downstream pull when the upstream finishes.