add support for a dark theme
Adds a theme parameter that can be used with graph
images. There are two themes right now: `light` and
`dark`. The `light` theme is the default and renders
the same way as before.

The default armytage palette is intended for a white
background. This change also adds two new palettes
from the [R Polychrome] package. The `light24`palette
is the default when using the `dark` theme.

[R Polychrome]: https://rdrr.io/rforge/Polychrome/man/palettes.html

This change is still experimental and may change prior
to 1.7.0 release.