add cluster group by operator
Maintains a sub-flow for each member of a cluster. The set
of flows can change as the membership of the cluster changes.
This is a more generic version of the connection manager
logic from the eval library.

Primary use-case is for operations that require communication
with all members of a cluster.