update lwcapi to accept `step` (#1125)
The goal is to try and make the naming more consistent
and standardize on `step` everywhere. Parts of lwcapi
were the only places that referred to this as `frequency`.

This updates the `ExpressionMetadata` class so it will
treat `step` as an alias. This will allow the eval library
to be updated. For now the output of the expressions api
needs to keep it as `frequency` to ensure backwards
compatibility with clients.