atlas-eval: fix eval flow failure on lwc instance removal
Currently the whole eval flow may fail when a lwc instance is removed.
There's a subflow for lwc api call of each lwc instance, and that subflow is wrapped around by "RestartFlow" to enable auto-retry.
The subflow is finished from upstream when a lwc instance is removed, and upstream of RestartFlow is finished, RestartFlow
finishes and propagate the failure from the current ongoing attempt if any, to downstream, and eventually propagate to main flow.

The fix is to have downstream of RestartFlow suppress the failure, since the failure is expected and can be ignored on lwc instance removal.