add support for rewrite based on host name
In some cases we have multiple DNS names for the same
backend, but the behavior varies a bit depending on the
host name. Internally this has been done for quite a
while on the main backends, but was not done for the
streaming execution. This change adds support for host
based rewrite that will be consistent between streaming
and the main backends.