freeze should reset to initial variables
The freeze operation is meant to reset the state of the
interpreter to isolate expressions. Before it was clearing
the variables meaning that after a freeze operation any
variables passed in at the start of the execution would
not be available. This change fixes it to reset to the
initial set of variables used when starting the execution.