eval: use BroadcastHub instead of Publisher
Updates the queue source used for writing diagnostic
messages to use a BroadcastHub instead of Publisher.
The main reason is that under newer versions of akka
there is a 5s subscription timeout for the underlying
FanoutProcessor created with the Publisher.

All other users of `asPublisher` were also converted
unless the `Publisher` type was explicitly needed as
part of the public API for the method.