improve hash function for int collections
For some data sets we are seeing it take a long time
to rebuild the index with most of the time being spent
doing get/put for the `RefIntHashMap`s used. Testing
with an improved int hash reduces collisions providing
a significant improvement.

Table showing reduction in build time using the newer
hash:

| Index size | Before | After  |
|------------|--------|--------|
| 1M         |    11s |    11s |
| 2M         |  1m43s |    20s |
| 5M         | 12m34s |    55s |