QueryIndex: reduce iteration cost for tags map
In #1225, we switched to using the EntriesIterator for
`SmallHashMap`s to avoid allocations. This works well for
compact maps. However, when there is free space in the map
it adds some computational overhead to search for the next
entry. Since the traversal occurs multiple times when
searching for matches this overhead can add up. In this
change we switch to a simple array with thread local so
we can iterate the tags map a single time and avoid the
allocations for most calls.