add helper for accessing opportunistic EC
In scala 2.13.4 the behavior of the default global execution
context changed. See:

https://github.com/scala/scala/releases/tag/v2.13.4

This change adds a helper class for accessing the opportunistic
context that was used before and switches some use-cases to
use that version as it is more well tested on our use-cases.
We can reevaluate in the future as needed.