trace logging for request exceptions
Add a log message that can be enabled to get the exception
coming from failed requests. In some cases there are
failures that show in the access log that are not otherwise
logged. This ensures that any exception leading to a failed
response can easily be logged via the request handler.