binary ops with a subset of group by keys
This change add support for binary operations where both
sides are grouped and the keys for one side is a subset
of the other. Before the keys would have to match exactly.
When one side has a subset, the matching for entries will
be performed using the smaller set of keys.

For example, suppose that dimensions `a` and `b` each have
two values:

```
Expr: q1,(,a,b,),:by, q2,(,a,),:by, :add

Result Set:
      q1:a1b1 + q2:a1
      q1:a1b2 + q2:a1
      q1:a2b1 + q2:a2
      q1:a2b2 + q2:a2
```