add topk/bottomk that aggregates others
Fixes #1224.

Adds variants of topk/bottomk that in addition to computing
the highest priority values will also return an aggregate
time series that includes all of the other time series that
were not high enough priority. This can be useful to see
what proportion of the overall volume is represented by
the highest priority time series. The operators have the
same signature as `:topk` and `:bottomk`, but the
aggregation to use is specified in the operator name.
There are four aggregates supported:

- min
- max
- sum
- avg

Count was not included as an aggregate because it could be
ambiguous and does not seem useful.

As part of this work there was some refactoring of the
existing aggregation operations to make the count case
more consistent.