switch blockingQueue to use BoundedSourceQueue (#1260)
There are some race conditions in the implementation of
the custom blocking queue source that result in:

- The source hanging after complete.
- An item in the queue not getting processed until another
  item is offered if the source is pulled after the push
  immediately check and before the poll call.

In recent versions of akka there is now a bounded source
queue that addresses the memory concerns we had. Rather
than track down the race conditions, this change switches
the internals of the `blockingQueue` helper to use the
BoundedSourceQueue in akka.