ensure alpha setting is not lost with toString
When processing operators that modify colors the later
operations will override and win. The expression string
generated for StyleExpr was sorting which means that
`:alpha` would be output before `:palette` even if the
user did them in the reverse order. Since `:palette`
will override a previously set alpha this caused the
setting to get ignored. For now change the sort order
to be descending which should always preserve the user
intent with the current set of style operations.