add directive to force connection close
For load balancers that work at the TCP level there can
be an imbalance in traffic if clients reuse connections.
The `closeConnection` directive can be used to mitigate
this by adding an explicit `Connection: close` header to
the responses.

In order to allow some reuse, a probability can be set
to tune how frequently the header is applied. This also
avoids the need to track details like the number of
requests per connection for more complex policies.