ensure normalized values are output in order (#1297)
Before the sum and max value functions would output the
current aggregate for a given time on each update. This
means the downstream would receive the values out of order
if that is how the values came in. In some cases, like
the hour transition for blocks, this can trigger a roll
over and prevent older data.

Now, those functions will maintain a buffer and delay the
output until either new data triggers a flush for the old
interval or there hasn't been an update and the entry
expires. This ensures the data gets output in the correct
time order. However, it does mean the data would appear
to be a bit more delayed on arrival since we have to wait
for it to complete.

The `LastValueFunction` was no longer in use and has been
removed.