add compressed mutable block type
Adds a new block type that is mutable, allows random access,
and tries to compress data as it is coming in. This should
reduce the overhead for the data that is filling in and help
avoid issue for out of order updates since the historical
blocks will not be compressed to an immutable type.