allow out of order updates for blocks
In #1297, normalized values were buffered and would only
get output at the end. This led to some user complaints
that data was delayed. It was still within SLA, but it
wouldn't show up until a bit later than before if the end
time was set to now instead of offset by the SLA for data
to be actionable.

With the compressed mutable block type introduced in #1320,
it is possible to update the older blocks as they are not
immutable. In this change, the block store checks have been
updated so the previous block can be updated to account for
out of order values that may arrive. This should make the
data visible earlier as it was before.