add SortedTagMap implementation
For some of the storage experiments it is useful to be
able to receive and maintain the tags sorted by the keys.
This change adds an implementation of Map that is based
on a sorted string array. The memory overhead and general
performance is similar to SmallHashMap, though lookups for
an individual key can be slower, logarithmic rather than
typically constant, compared to a non-compact small hash
map instance.