add simplified tag index (#1343)
Experiment for a more restrictive tag index that only
supports finding items overall and not projecting keys
or values. The index is also setup so that it doesn't
need to reference the items or tag maps to give more
flexibility to the database implementation over how that
data is stored and structured.

The implementation is heavily based on the corresponding
parts of the RoaringTagIndex implementation. If the overall
experiments go well, then this will replace the various
implementations of TagIndex.