reduce overhead of expressions etag check
Before it would compute the etag by computing a sha1 hash
of a string representation of the sorted expression list.
With 100k+ expressions that is quite expensive both in
terms of the memory use and CPU.

This change introduces a cache for the encoded entity so
it can be returned directly if there are frequent calls
within a short period. The ETag is computed over the
encoded and compressed response using CRC32. This reduces
the memory and CPU overhead considerably.