optimize validation when using SortedTagMap
Add validate variant that allows checking this type without
first converting to SmallHashMap.

Benchmark was updated and fixed so that validation would pass
which is the most common case in actual use. Before it would
fail due to the reserved key check. That made the benchmark
results misleading as the composite check seemed faster since
it was able to short-circuit earlier. For the successful case
it was actually slower. Also adds a flag that can be used to
choose whether or not to group simple tag rules into a
composite. By default it will not as that seems to generally
be faster.