use array instead of int map for itemTags
The itemTags field of the index is only used for findKeys
and findValues. In the findKeys case it loops over all keys
and the simple array is faster. For the findValues case, it
gets the value for a key which is now linear for the number
of tags so will be a bit slower.

The main benefit is it can reduce the memory use considerably
for large indexes. The IntIntHashMap has an overhead of 40
bytes and also allocates an array of 2x the size to avoid
collisions in the map. With a sample index of 10M items this
change reduced the memory size by several GB.