support custom palettes in expr (#1358)
Update the custom palette format to be a list of colors
that can be used either in the URI or directly in an expr
with the `:palette` operator. For example:

```
Via URL query parameter:

/api/v1/graph?q=
    name,sps,:eq,(,nf.cluster,),:by,:stack
  &palette=(,1a9850,91cf60,d9ef8b,fee08b,fc8d59,d73027,)

Via expr:

/api/v1/graph?q=
    name,sps,:eq,(,nf.cluster,),:by,:stack,
    (,1a9850,91cf60,d9ef8b,fee08b,fc8d59,d73027,),:palette
```

The old URL format with `colors:` prefix will still work
via the URL parameter for backwards compatibility.