support for renaming tag key (#1366)
Adds `:as` operation that can be used to rename a tag
key from an expression. The primary use-case is being
able to perform binary operations where the corresponding
key has a different name on the each side. One side can
now be renamed to have them match.

For example, with common IPC metrics the server app on
the server metrics would be `nf.app`. The server app on
the client metrics would be `ipc.server.app`.

Fixes #789.