add codec for more efficient publish format (#1379)
Adds helper functions for a new publish format that is more
efficient to decode. In particular, it allows the id to be
sent along so it doesn't need to get recomputed and keeps the
strings up front so they are only interned once if there are
duplicate values.

This format is a bit more expensive to encode due to needing
to compute the string table. This change also Moves the helper
functions for encoding and decoding publish payloads to a
separate object.