incremental expression evaluation
Add method for TimeSeriesExpr to get a evaluator that can
be used to incrementally process a set of time series
without needing to have the entire set in memory. This
depends on being able to read the data in a given order
so that expressions can be processed and ensure a grouped
output time series is complete without needing to wait
for the entire dataset to be processed.