support stripping out style and filter operations
Updates the `/expr/strip` API to allow filtering operations
from the style and filter vocabularies. This can be useful
for some use-cases for ignoring presentation aspects and
trying to see if the underlying expression is the same. In
the future other operators maybe supported as well, but that
becomes more complex as it wouldn't necessarily be a single
expression.