lwc: use asg for matching exprs to a cluster (#1431)
Some expressions have an exact match for `nf.asg`, but do
not have conditions for `nf.app` or `nf.cluster`. This change
will parse the asg value based on the server group naming
rules and add a cluster restriction so that expressions for
a specific asg will only show when requesting data for the
cluster containing that asg.