fix id when common tags are after metric data (#1439)
If the common tags were after the metric data in the batch
json payload, then the tags in the tuple would get updated,
but not the id. This change fixes it so that the id is
updated based on the final tags for a metric. Also updates
the behavior so the conflicts resolution is consistent for
both paths.