fix common grouping for aggr helpers (#1440)
Updates the rewrite used for math functions that behave
like aggregation functions, e.g. `:avg`, so that the
rewrite works correctly when the display expression is
a `Query`. Before it would always apply the rewrite to
just the eval expression which could lead to a different
and sometimes broken expression if reinterpreted from
the display expression.