fix cg with custom averages (#1444)
The subset check for nested grouping would check all
data expressions. If the underlying expression was
combined with binary operations with the grouping for
one side being a subset of the other, then a subsequent
grouping operation could fail the check. Now it checks
based on the final grouping of the pre-aggregate.