1.6: fix shard lookup for Integer.MIN_VALUE (#1459)
Back port of #1452.

Before it was relying on `abs` to get a non-negative value
for computing an index to the array. This breaks when the
int value of the id is `Integer.MIN_VALUE`. Update it to
explicitly clear the sign bit to ensure the value will
always be non-negative.