adjust consolidation behavior for rates (#1467)
Port an old internal fix to ensure we have consistent
behavior. If the data source type is a rate, then treat
NaN values as 0 as it means no activity measured for
those intervals when consolidating. That means the avg
will always be computed over the entire consolidated
interval.

For gauges, the NaN values are ignored because to be
consistent with values reported to the gauge. The
behavior of gauges is not impacted with this change.