optimize result tags computation (#1469)
A long time ago the tag set was made deterministic
based on the query expression where it only includes
tags with an exact match or that are listed in the
group by. As a result, it isn't necessary to compute
the intersection of dimensions while aggregating the
matching time series.