allow image to be disabled for v2.json (#1476)
The image rendering can be computationally expensive
and increase the size of the payload quite a bit. For
some cases it is desirable to use the `v2.json` format
since it is similar to fetch and LWC, but the image is
not needed. Alert processing is one example.

This change adds a `hints` URI parameter that can be
used to pass arbitrary hints back to the rendering engine
that is used. The `v2.json` engine has been updated to
omit the image if a `no-image` hint is passed. In the
future we may want to only include the image if requested,
but that would be backwards incompatible and break some
current usage.