use custom JSON encoding if specified (#1477)
If a class extended JsonSupport and provided a custom
encoding it would get used by the interited methods, but
would get ignored when using `Json.encode`. This could
create problems when an object is nested because the
custom encode would not get used. The `Json.encode`
could not directly call encode, because it would create
a loop with the default implementation.

This change adds a method that can indicate if it is
desirable for `Json.ecnode` to use the custom encoding.