json: fix double parsing for large integers (#1478)
Before when using `JsonParserHelper.nextDouble` if the raw
encoded value looked like an integer type it would try to
access it as a long, then convert the long to a double.
This causes problems for larger integers that cannot fit
within the bounds of a Java long. Now it will access it
as a double.