improve performance of batch index update
A lot of the time and allocations were being spent sorting
the items passed to the index. Java uses TimSort internally
when sorting object arrays and that uses a merge sort. It
allocates a bunch of temporary arrays when performing the
merges.

This change leverages the fact that the items that are
already indexed will be sorted, so we can just sort the
new items and merge the two sorted sets.