lwc: keep input to TimeGrouped batched
Refactors the stream from parsing the messages from lwcapi
to the TimeGrouped stage to keep the datapoints batched.
This significantly reduces the number of messages passed
between the stream stages and achieve a higher throughput.

Also creates a dedicated thread pool to use for parsing
the messages as that is the most time consuming step.