add hint for controlling y-axis color (#1492)
Adds a rendering hint, `ambiguous-multi-y`, for the
graph image that will disable the behavior of using
the first line on an axis for the axis color. Instead
it will use the color selected based on the theme.

As indicated by the name, in this mode there will not
be any visual cue as to which axis corresponds to
data on the graph image.