add value function to dedup inputs (#1500)
When pushing data to memory the blocks can easily overwrite
when an update for a given datapoint is received. In other
use-cases this can be quite expensive and undesirable. This
change adds a value function that can be used to buffer the
updates for a window of time and dedup the data.