remove redundant clauses when normalizing
In some cases with tooling a `:cq` can get applied multiple
times. If the common query condition has an OR clause, then
that will get expanded with the cross product when computing
the DNF. As a result, the normalized query resulting from
applying a `:cq` once would differ from applying the same
`:cq` multiple times. This change removes redundant clauses
from the normalized OR set to avoid the discrepancy. Redundant
means that the removed clause will not change the result
since another branch of the OR would have matched.