eval: sum aggregation for gauges (#1534)
Adds a new datapoint aggregator implementation to handle
sum/count of gauge values that come from an aggregation
service where there can be duplicates. The aggregation
service would add a tag `atlas.aggr` with a key to use
for grouping and deduping the relevant values. Values
within a group will use max aggregation to be consistent
with the behavior if there was a single instance of the
aggregation service.

Backport of #1533.