lwc: model object for event
Initial model object representing an arbitrary event to process
with LWC. The raw event would typically be a log, span, or trace.
The `LwcEvent` type would define how to extract data and encode
the event without putting any restrictions on the structure so
that LWC is agnostic to the event model and so we can avoid
expensive conversions when processing.