lwc: add timestamp accessor
When mapping to time series data points, we need to know
the timestamp for an event. For events like spans that
have a start and end, the end time is used. This is similar
behvavior to a timer which would get updated at the end when
the latency is known.