add small set of named colors
The main use-case is where a particular color is desirable
to indicate an intent. For example, using green for success
and reds for failures. When using themes, explicit colors
can be problematic when trying to ensure good contrast with
the background which could vary depending on the theme picked
by the user. The exact color for the named set will vary for
a given theme to try and have good contrast.