add support for heatmaps (#1547)
Adds a heatmap line style that can be used to
render a set of lines on an axis as a heatmap
instead of a line. There can only be a single
heatmap on a particular axis. The `:heatmap`
operator can be used to select this style.

The cells for the heatmap are determined based
on the step size for the time axis and the tick
grid for the value axis. For normal lines each
line that overlaps a given cell will contribute
one to the count.

For percentile approximations that would used
with the `:percentiles` operator, they can now
be used with `:percentiles-heatmap` to map the
distribution to a heatmap. In this case the
count for a cell will be determined by the
distribution for the percentile approximation.

Additional axis specific parameters have been
added for `scale`, `u`, `l`, `palette`, and
`label` with a prefix of `heatmap_`. The scale
will control how counts are mapped to colors.

Since a heatmap can combine many lines, the
default legend test is just `Heatmap`. The
`heatmap_label` parameter on the URL can be
used to customize the label.