adjust heatmap scale behavior (#1551)
Changes the min to be the minimum non-zero count for
a cell, or zero if there is no data. Updates the color
scale to use the actual min/max or the bounds. Before
it was using zero even if an explicit lower bound was
selected. Further, using the min non-zero count and
max count as the range spreads the colors better over
the data range.

This also removes the max behavior of rounding to
one significant digit. That rounding would lead to tick
values that were easier to reason about, but it could
round by a large amount relative to the data causing
most data to go to a single color cell. Will revisit
in the future for achieving more significant tick
boundaries with less downsides.

Some minor fixes to the inverted scale logic for the
log-linear scale.