eval: ensure exprs are resent on reconnect (#1557)
If the websocket to an lwcapi node fails and needs to reconnect,
then the message with the set of expressions needs to be resent.
Before this would only happen when there was a change in either
the set of remote instances or the set of expressions. With this
change, the set of expressions will get repeated every 5s so it
will get resent on a reconnect.

Note, the websocket sub-flow already has some deduping if there
is no change, so the additional overhead during normal operations
is minimal. If there is a failure, then the sub-flow will get
restarted and the state for the `unique` operator will get reset
allowing the repeated expression message through.