permit all allowed step sizes when computing (#1573)
Before there was a separate set of auto step sizes that was
a subset of the allowed step sizes. This can sometimes mean
the jump is larger than some users would like. The original
reason for this was to ensure a higher hit rate for an old
caching setup, but is no longer relevant.