core: add sample-count operation
Adds a `sample-count` operation that can be used to estimate
the number of samples received for a percentile approximation
baed on a range of the distribution. This estimate is computed
by adjusting the query to restrict the set of buckets used to
model the distribution to those for the specified range.

One use-case for this is to be able to track number of requests
that are not within an SLO based on latency. If more precise
tracking is needed, then a bucket counter could be used with a
bucket boundary that matches the SLO threshold. However, this
operation can be used to get an estimate if an existing
percentile timer exists.