core: parse epoch in micros or nanos
Updates the data parsing to allow for epoch timestamps
in seconds, millis, micros, or nanos. Before it only
supported seconds and millis. This does reduce the range
of times a bit, but in practice the transition point is
in the 1970s well before times we encounter or a few
hundred years in the future.