pekko: source based on BlockingQueue (#1676)
Add helper to create a source that is based on a Java
BlockingQueue implementation. This can be used as an
alternative to Pekko's BoundedSourceQueue when more
control of the queue is desirable. If the queue is full
it can be set to either drop newly arriving values or
older values that are already in the queue.