pekko: update client helper to support flow API
Update PekkoHttpClient to support `superPool` more commonly
used with streams. Also hooks in retry logic to make it work
with the logger consistently and support throttling. Pekko's
internal pool retries only happen if the request fails with
an exception. This also provides some more flexiblity with
the retry decision so it may be used for requests that are
not clearly idempotent based on the HTTP spec, but maybe ok
to retry for a given use-case.