lwcapi: avoid queue for high priority messages
Some messages such as the subscription metadata need to go
through in order to properly process data. In some cases we
would see a lot of input data come in almost immediately and
result in the subscription info messages getting dropped. The
consumer was then unable to match the datapoints with an expr.

This changes the messages for setting up a new subscription to
be a high priority path that bypasses the queue for incoming
data points.