eval: add a rewrite module to handle queries with namespaces (#1675)
If configured via `atlas.eval.stream.rewrite-url`, data source URIs are sent to a
rewriting service. Responses with errors are then dropped and the caller notified.
Valid rewrites continue processing.

Some uses, e.g. the atlas-stream case, entail updating the data sources for a
flow when new subscriptions are added or existing subscriptions dropped.
This will return cached rewrites for individual queries in the event the
rewrite backend is unavailable. Calls will continue for the rewrite service
until a successful response is received, at which point it will update the
cache and forward the rewrites downstream.