core: remove cached hash code for ItemId (#1696)
The id is generated from a hash function, typically SHA1 so
we can just use a part of that for the hash code of the id
object. Since data is sometimes partitioned and routed based
on the most and least significant bits, we pick one of the
middle bytes to reduce the liklihood of collisions.

Benchmarks show it is about 25% slower to access the hash
code this way, but it does provide some memory savings and
compute savings when creating an id instance since the murmur
hash doesn't need to be computed.