add event sample operator
This operator allows events to be sampled based on a set
of tag keys. It works similar to an analytics group by,
but will capture some projection fields for a sample event
to provide more context to the consumer.

This change also adds a max groups limit for the LWC events
client. The limit is needed to help avoid excessive memory
usage if the sampling keys have a high cardinality.