core: allow unix timestamp to be anchor time parameters
Updates the date and range parsing to allow a unix timestamp
value to be an absolute time anchor for the time parameters.
This is mostly a convenience if generating the parameters via
substitution, but there is a need to have some offset.

For now the ISO date time patterns do not support the offset
behavior. The parsing is trickier because they already use
`-` for the year-month-day as well as for time zone offsets.