core: fix corner cases with less than comparison
For the roaring tag index there were some bugs resulting
in incorrect matching if the value being compared was larger
than the last element for a given tag. The offset position
was for the last tag and so it would return an empty match
set since the tag key didn't match.