core: deterministic tagging for aggregated groups (#1749)
Updates aggregation functions applied to the result of a
group by to have a deterministic set of tags based on the
exact matches. Before if there was a single group, then
the group tag would be preserved in the result causing the
tag set for the result to vary depending on the data that
matched.