core: improve handling of large spikes for rolling-sum (#1753)
Update rolling-sum and rolling-mean to better handle cases
where there is a large value. Before a really large value
compared to others would skew the sum and it would never
fully recover since the sum is maintained incrementally. Now
it will check if there is a big difference in the exponent
and recompute from the buffer to recover.