core: fix pagination and lazily compute OR
Update RoaringTagIndex implementation to fix pagination
of findItems for range and pattern clauses. Before the
offset was being ignored for these query clauses. In
common usage it often wouldn't be noticed because AND
with an equal clause would result in the offset gettng
applied via the conjunction.

Also updates range queries and patterns to use a lazy
OR operation on the bitmap. In some of our benchmarks
this improves the throughput by about 30%.