eval: optimize computing list of datapoints
For the GroupByAggregator optimize computing the datapoints.
Use a List builder to avoid creation of iterators and the
intermediate collections when doing flatMap and then converting
back to a List. Since all aggregators that are part of a group
should have a single element in the list, special case that
path to avoid the need for a separate iterator.