lwcapi: update expression cache behavior (#1777)
Before it would cache for 10s and then re-encode the next
time a key was fetched. This could result in many requests
waiting for the encode to complete on the next access after
expiration. Now it will always return the cached copy if
available and refresh them in the background.