lwcapi: add option to ignore publish step (#1778)
When listing expressions for a cluster, the client will
often ignore the publish step because it is covered by
a bridge cluster. Add an option so we can disable listing
those for a cluster to reduce the payload sizes.

Also fixes updates so that refreshing will not update the
access time for the cache entry.