eval: support config based rewrite (#1799)
Add support for a simple config based mapping to allow
for a generic namespace query layer with `ns` query param
to rewrite to supported hosts. This replaces the http
approach since it is a simple rewrite to avoid the risk
of problems with that service impacting streaming execution
of data.