core: move `:cq` to query vocabulary (#1800)
This allows it to be used for APIs that only accept a query.
The main benefit is to support `:list,(,contextQuery,:cq,),:each`
pattern. It could be done with as `baseQuery,contextQuery,:and`,
but in some cases it is convenient for tools to use the same
pattern for tags and graph API.