Updating graph engine to apply constraints
It was found that the userError and systemError were not utilizing the GraphConstants.MaxWidth and GraphConstants.MaxHeight. Applying a validation step in the createImage of the DefaultGraphEngine so that all requests receive this clamp of values will help reduce the chance of a out of memory error. This applies to requests with valid or invalid queries that have dimensions that are out of bounds for the GraphConstants values defined.